+++ backend/src/main/java/com/pickgo/domain/area/area/entity/PerformanceArea.java
@@ -1,11 +1,11 @@
-package com.pickgo.performance.performance.entity;
+package com.pickgo.domain.area.area.entity;
 
 import com.pickgo.global.entity.BaseEntity;
-import com.pickgo.performance.seat.entity.Seat;
+import com.pickgo.domain.performance.entity.Performance;
+import com.pickgo.domain.area.seat.entity.Seat;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,19 +15,22 @@ import java.util.List;
 @Getter
 @Setter
 @Builder
-public class PerformanceSession extends BaseEntity {
+public class PerformanceArea extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private LocalDateTime performanceTime;
+    private String name;
+
+    @Column(nullable = false)
+    private Integer price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "performance_id", nullable = false)
     private Performance performance;
 
     @Builder.Default
-    @OneToMany(mappedBy = "performanceSession", cascade = CascadeType.ALL, orphanRemoval = true)
+    @OneToMany(mappedBy = "performanceArea", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Seat> seats = new ArrayList<>();
 }
