+++ backend/src/main/java/com/pickgo/domain/area/area/repository/PerformanceAreaRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.performance.area.repository;
+package com.pickgo.domain.area.area.repository;
 
-import com.pickgo.performance.area.entity.PerformanceArea;
+import com.pickgo.domain.area.area.entity.PerformanceArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PerformanceAreaRepository extends JpaRepository<PerformanceArea, Long> {
