+++ backend/src/main/java/com/pickgo/domain/area/seat/entity/Seat.java
@@ -1,8 +1,8 @@
-package com.pickgo.performance.seat.entity;
+package com.pickgo.domain.area.seat.entity;
 
 import com.pickgo.global.entity.BaseEntity;
-import com.pickgo.performance.area.entity.PerformanceArea;
-import com.pickgo.performance.performance.entity.PerformanceSession;
+import com.pickgo.domain.area.area.entity.PerformanceArea;
+import com.pickgo.domain.performance.entity.PerformanceSession;
 import jakarta.persistence.*;
 import lombok.*;
