+++ backend/src/main/java/com/pickgo/domain/area/seat/repository/SeatRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.performance.seat.repository;
+package com.pickgo.domain.area.seat.repository;
 
-import com.pickgo.performance.seat.entity.Seat;
+import com.pickgo.domain.area.seat.entity.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SeatRepository extends JpaRepository<Seat, Long> {
