+++ backend/src/main/java/com/pickgo/domain/performance/entity/PerformanceSession.java
@@ -1,11 +1,11 @@
-package com.pickgo.performance.area.entity;
+package com.pickgo.domain.performance.entity;
 
 import com.pickgo.global.entity.BaseEntity;
-import com.pickgo.performance.performance.entity.Performance;
-import com.pickgo.performance.seat.entity.Seat;
+import com.pickgo.domain.area.seat.entity.Seat;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,22 +15,19 @@ import java.util.List;
 @Getter
 @Setter
 @Builder
-public class PerformanceArea extends BaseEntity {
+public class PerformanceSession extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private Integer price;
+    private LocalDateTime performanceTime;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "performance_id", nullable = false)
     private Performance performance;
 
     @Builder.Default
-    @OneToMany(mappedBy = "performanceArea", cascade = CascadeType.ALL, orphanRemoval = true)
+    @OneToMany(mappedBy = "performanceSession", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Seat> seats = new ArrayList<>();
 }
