+++ backend/src/main/java/com/pickgo/domain/performance/repository/PerformanceRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.performance.performance.repository;
+package com.pickgo.domain.performance.repository;
 
-import com.pickgo.performance.performance.entity.Performance;
+import com.pickgo.domain.performance.entity.Performance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PerformanceRepository extends JpaRepository<Performance, Long> {
