+++ backend/src/main/java/com/pickgo/domain/performance/repository/PerformanceSessionRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.performance.performance.repository;
+package com.pickgo.domain.performance.repository;
 
-import com.pickgo.performance.performance.entity.PerformanceSession;
+import com.pickgo.domain.performance.entity.PerformanceSession;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PerformanceSessionRepository extends JpaRepository<PerformanceSession, Long> {
