+++ backend/src/main/java/com/pickgo/domain/performance/repository/VenueRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.performance.performance.repository;
+package com.pickgo.domain.performance.repository;
 
-import com.pickgo.performance.performance.entity.Venue;
+import com.pickgo.domain.performance.entity.Venue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VenueRepository extends JpaRepository<Venue, Long> {
